Add top-level expressions
"Top-level expressions" are expressions like `def` and `typ` that occur
first in the file, and can be preceeded only by other top-level
expressions. They can be parsed faster because they are of form

  (keyword [stuff]) rest

where `rest` is of type `Exp`, and so a fold can be used to fuse these
expressions together with whatever occurs after the last top-level
expression. This is significantly faster and uses much less memory
because the `repeat` parser combinator can be used, rather than explicit
recursion on every single top-level expression.